+++ src/components/layouts/RightBar.tsx
@@ -1,7 +1,7 @@
 import { Box, Button, HStack, Text, VStack } from '@chakra-ui/react'
 import Link from 'next/link'
 import { ReactNode } from 'react'
-import { SPONSERS } from '../../constants/sponser'
+import { SPONSERS } from '../../entities/Sponser'
 import QAButton from '../atoms/QAButton'
 import { SponserBanner } from '../common/suponser/Banner'
 import { RIGHT_BAR_HEIGHT, RIGHT_BAR_MINI_WIDTH, RIGHT_BAR_WIDTH } from './constants'
@@ -15,7 +15,7 @@ export const RightBar = ({ children }: Props): JSX.Element => {
         <VStack as="aside" bgColor="#FBF6F0" h={RIGHT_BAR_HEIGHT} w={RIGHT_BAR_WIDTH} minW={RIGHT_BAR_MINI_WIDTH}>
             <VStack spacing="15px" margin="30px 10px">
                 <SponserBanner sponser={SPONSERS.tasuki} />
-                <SponserBanner sponser={SPONSERS.shinnichi} rate={0.7} />
+                <SponserBanner sponser={SPONSERS.shinnichi} />
             </VStack>
             <QAButton />
         </VStack>
