+++ src/components/molecules/profile/EditForm.tsx
@@ -1,9 +1,9 @@
 import { FormControl, FormErrorMessage, Input, Select, Checkbox, VStack, FormLabel } from '@chakra-ui/react'
-import Department from '../../../entities/department'
+import Department from '../../../entities/Department'
 import React from 'react'
 import { User } from '../../../entities/user'
 import { PrimaryButton } from '../../common/buttons/PrimaryButton'
-import { gradeList } from '../../../entities/grade'
+import { gradeList } from '../../../entities/Grade'
 import { useCreateUserForm } from '../../../hooks/useCreateUserForm'
 import { UserReq } from '../../../api/req/UserReq'
