+++ src/components/pages/Register/Step3/Page.tsx
@@ -15,8 +15,8 @@ import {
 import { useRouter } from 'next/router'
 import { PrimaryButton } from '../../../common/buttons/PrimaryButton'
 import { useCreateUserForm } from '../../../../hooks/useCreateUserForm'
-import Department from '../../../../entities/department'
-import { gradeList } from '../../../../entities/grade'
+import Department from '../../../../entities/Department'
+import { gradeList } from '../../../../entities/Grade'
 import { useErrorToast } from '../../../../hooks/errors/useErrorToast'
 import { ERROR_MESSAGE } from '../../../../constants/errors'
 import { PAGE_LINKS } from '../../../../constants/pageLinks'
